Op.62 No.1 proofreading
Measure 5, beat 1, right hand: Add missing cross-voice arpeggio. Since
this change affected spacing for the entire line, the rendering of the
manual slur from measure 1 (\slurShapeA) needed a slight adjustment.

Measure 22/24/25, beats 3 and 4, left hand: The notes on these beats in
leftHandLower should be B-naturals instead of A-sharps.

Measure 59, beat 2, left hand: The lowest note of the second chord
should be F-natural instead of E-flat.

Measure 76, beat 2, left hand: The C-natural in the upper voice should
be a dotted-quarter note instead of a quarter note.