Op.62 No.1: proofreading slurs and pedals
Measure 26, beat 3, right hand: Add missing slur to the fioritura line,
extending to the dotted-half note in the following measure. Note that in
the source edition, the slur appears above the beam, but using a \slurUp
here would then require creating a custom slur shape. I will avoid that
for now, since the appearance of the default slur is probably
acceptable.

Measure 66, beats 1 and 2: Move the pedal release mark closer to beat 3
to align with harmony changes.

Measure 69, beat 1: Add missing pedal engage/release marks for the first
eighth-note duration. Also adjust remaining pedal markings in this
measure to appropriately stagger them.

Measure 71, beat 1: Remove superflous pedal marking on the second half
of this beat. Instead, the pedal that begins at the beginning of the
measure should be extended for a full beat.