Op.62 No.2: proofreading slurs and pedals
Measure 53, beat 1: Change pedaling to the match that from the
corresponding beat in the following measure, since the harmony and
figurations are identical there as well.

Measure 61, beats 3 and 4, right hand: Adjust endpoint of the slur that
began in measure 58.

Measure 69, beat 2: Change pedal release to match that from beat 4 of
this measure.